Refactor Python Data Source to aline with other built-in Data Sources
